requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aiosignal==1.3.1
albucore==0.0.23
albumentations==2.0.4
altgraph==0.17.4
annotated-types==0.7.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==2.0.12
click==7.1.2
colorama==0.4.6
contourpy==1.2.1
cryptography==43.0.0
cssselect==1.2.0
cycler==0.12.1
Cython==3.0.12
defusedxml==0.7.1
dirsearch==0.4.3
easydict==1.13
environs==9.5.0
et_xmlfile==2.0.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
fake-headers==1.0.2
Faker==26.3.0
filelock==3.16.0
Flask==1.1.4
flatbuffers==24.12.23
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.9.0
gast==0.6.0
gevent==23.9.1
gmpy2==2.2.1
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.66.1
h11==0.14.0
h5py==3.12.1
html5lib==1.1
idna==3.4
imageio==2.37.0
importlib_metadata==8.4.0
importlib_resources==6.4.0
insightface==0.7.3
itsdangerous==0.24
Jinja2==3.1.4
joblib==1.4.2
JsonForm==0.0.2
jsonpath==0.82.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
JsonSir==0.0.2
keras==3.7.0
kiwisolver==1.4.5
lazy_loader==0.4
legacy==0.1.7
libclang==18.1.1
loguru==0.7.2
lpips==0.1.4
lxml==4.9.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.8.4
maxminddb==2.6.2
maxminddb-geolite2==2018.703
mdurl==0.1.2
ml-dtypes==0.4.1
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.0.5
mysql-connector-python==9.0.0
namex==0.0.8
networkx==3.2.1
ntlm-auth==1.5.0
numpy==1.26.4
onnx==1.17.0
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.13.1
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.3
pefile==2024.8.26
pillow==10.3.0
prettytable==3.14.0
protobuf==5.28.1
psycopg==3.2.1
psycopg-binary==3.2.1
PyAutoGUI==0.9.54
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.10.6
pydantic_core==2.27.2
pyDes==2.0.1
PyGetWindow==0.0.9
Pygments==2.18.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyMsgBox==1.0.9
pyOpenSSL==24.2.1
pyparsing==3.1.4
pyperclip==1.9.0
pyquery==1.4.3
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pyspnego==0.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
Python-EasyConfig==0.1.7
pytweening==1.2.0
pytz==2024.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
redis==3.5.3
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
requests_ntlm==1.3.0
Resource==0.2.1
retrying==1.3.4
rich==13.9.4
rpds-py==0.20.0
scikit-image==0.24.0
scikit-learn==1.5.0
scipy==1.13.0
seaborn==0.13.2
selenium==4.23.1
simsimd==6.2.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sspilib==0.1.0
stringzilla==3.12.1
supervisor==4.2.5
sympy==1.12
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.8.30
torch==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.67.1
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.12
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==1.0.1
win32-setctime==1.1.0
wordcloud==1.9.4
wrapt==1.17.0
wsproto==1.2.0
xlwt==1.3.0
yarl==1.9.4
zipp==3.18.1
zope.event==5.0
zope.interface==7.0.1
